perception/requirements.txt
fastapi
uvicorn
opencv-python
numpy
torch
torchvision
Pillow
onnxruntime
requests
timm
